got rid of some old code, brokoe out nav buttons to seperate component
