+++ portfolio/src/App.js
@@ -16,9 +16,9 @@ class App extends Component {
       <div id="main-intro"><p className = "intro-text"> Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
       </p> </div>
 
-      <div id='footer-wrapper'>
+      <footer>
         <Footer/>
-      </div>
+      </footer>
     </div>);
 
   }
